+++ src/graphic/Fast3D/Fast3dWindow.h
@@ -51,8 +51,8 @@ class Fast3dWindow : public Ship::Window {
     static bool KeyDown(int32_t scancode);
     static bool KeyUp(int32_t scancode);
     static void AllKeysUp(void);
-    static bool MouseButtonDown(int8_t button);
-    static bool MouseButtonUp(int8_t button);
+    static bool MouseButtonDown(int button);
+    static bool MouseButtonUp(int button);
     static void OnFullscreenChanged(bool isNowFullscreen);
 
   private:
